+++ entity/src/main/resources/db/migration/tables/user/V1.52__banned.sql
@@ -1,0 +1,5 @@
+-- ------------------------------------------------
+-- Version: v1.52
+-- -------------------------------------------------
+
+ALTER TABLE User ADD COLUMN banned BOOLEAN DEFAULT false;