+++ repository/src/main/java/org/grnet/cat/repositories/HistoryRepository.java
@@ -1,0 +1,8 @@
+package org.grnet.cat.repositories;
+
+import jakarta.enterprise.context.ApplicationScoped;
+import org.grnet.cat.entities.history.History;
+
+@ApplicationScoped
+public class HistoryRepository implements Repository<History, Long> {
+}